test: (#99) 구현체 테스트 분리
