--- src/test/java/org/dinosaur/foodbowl/domain/review/presentation/ReviewPhotoRepositoryTest.java
@@ -1,114 +1,0 @@
-package org.dinosaur.foodbowl.domain.review.presentation;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.SoftAssertions.assertSoftly;
-
-import java.util.List;
-import org.dinosaur.foodbowl.domain.photo.domain.Photo;
-import org.dinosaur.foodbowl.domain.review.domain.Review;
-import org.dinosaur.foodbowl.domain.review.domain.ReviewPhoto;
-import org.dinosaur.foodbowl.domain.review.persistence.ReviewPhotoRepository;
-import org.dinosaur.foodbowl.test.PersistenceTest;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-
-@SuppressWarnings("NonAsciiCharacters")
-class ReviewPhotoRepositoryTest extends PersistenceTest {
-
-    @Autowired
-    private ReviewPhotoRepository reviewPhotoRepository;
-
-    @Test
-    void 리뷰에_해당하는_리뷰_사진_엔티티를_조회한다() {
-        Review review = reviewTestPersister.builder().save();
-        Photo photo1 = photoTestPersister.builder().save();
-        Photo photo2 = photoTestPersister.builder().save();
-        ReviewPhoto reviewPhoto1 = ReviewPhoto.builder()
-                .review(review)
-                .photo(photo1)
-                .build();
-        ReviewPhoto reviewPhoto2 = ReviewPhoto.builder()
-                .review(review)
-                .photo(photo2)
-                .build();
-        reviewPhotoRepository.save(reviewPhoto1);
-        reviewPhotoRepository.save(reviewPhoto2);
-
-        List<ReviewPhoto> reviewPhotos = reviewPhotoRepository.findAllByReview(review);
-
-        assertThat(reviewPhotos).contains(reviewPhoto1, reviewPhoto2);
-    }
-
-    @Test
-    void 리뷰_사진_엔티티를_저장한다() {
-        Review review = reviewTestPersister.builder().save();
-        Photo photo = photoTestPersister.builder().save();
-        ReviewPhoto reviewPhoto = ReviewPhoto.builder()
-                .review(review)
-                .photo(photo)
-                .build();
-
-        ReviewPhoto savedReviewPhoto = reviewPhotoRepository.save(reviewPhoto);
-
-        assertSoftly(
-                softly -> {
-                    softly.assertThat(savedReviewPhoto.getId()).isNotNull();
-                    softly.assertThat(savedReviewPhoto.getReview()).isEqualTo(review);
-                    softly.assertThat(savedReviewPhoto.getPhoto()).isEqualTo(photo);
-                }
-        );
-    }
-
-    @Test
-    void 리뷰_사진_엔티티를_삭제한다() {
-        Review review = reviewTestPersister.builder().save();
-        Photo photo1 = photoTestPersister.builder().save();
-        Photo photo2 = photoTestPersister.builder().save();
-        ReviewPhoto reviewPhoto1 = ReviewPhoto.builder()
-                .review(review)
-                .photo(photo1)
-                .build();
-        ReviewPhoto reviewPhoto2 = ReviewPhoto.builder()
-                .review(review)
-                .photo(photo2)
-                .build();
-        reviewPhotoRepository.save(reviewPhoto1);
-        reviewPhotoRepository.save(reviewPhoto2);
-
-        long deleteCount = reviewPhotoRepository.deleteAllByReview(review);
-
-        assertThat(deleteCount).isEqualTo(2);
-    }
-
-    @Test
-    void 리뷰에_해당하는_사진_엔티티들을_삭제한다() {
-        Review review = reviewTestPersister.builder().save();
-        Photo deleteTargetPhoto1 = photoTestPersister.builder().save();
-        Photo deleteTargetPhoto2 = photoTestPersister.builder().save();
-        Photo photo = photoTestPersister.builder().save();
-        ReviewPhoto deleteReviewPhoto1 = ReviewPhoto.builder()
-                .review(review)
-                .photo(deleteTargetPhoto1)
-                .build();
-        ReviewPhoto deleteReviewPhoto2 = ReviewPhoto.builder()
-                .review(review)
-                .photo(deleteTargetPhoto2)
-                .build();
-        ReviewPhoto reviewPhoto = ReviewPhoto.builder()
-                .review(review)
-                .photo(photo)
-                .build();
-        reviewPhotoRepository.save(deleteReviewPhoto1);
-        reviewPhotoRepository.save(deleteReviewPhoto2);
-        reviewPhotoRepository.save(reviewPhoto);
-
-        List<Photo> deleteTargetPhotos = List.of(deleteTargetPhoto1, deleteTargetPhoto2);
-        long deleteCount = reviewPhotoRepository.deleteByReviewAndPhotos(review, deleteTargetPhotos);
-
-        assertSoftly(softly -> {
-            softly.assertThat(deleteCount).isEqualTo(deleteTargetPhotos.size());
-            softly.assertThat(reviewPhotoRepository.findAllByReview(review)).contains(reviewPhoto);
-            softly.assertThat(reviewPhotoRepository.findAllByReview(review)).doesNotContain(deleteReviewPhoto1, deleteReviewPhoto2);
-        });
-    }
-}
